refactor: changed the name of components and files to assure consistency with the rest of the project
